requirements.txt
requests
pandas
python-telegram-bot==13.7
flask
